Refactor DownloadManagerStats and GroupHeader components to import fragment types from the generated GraphQL file, improving type safety and consistency across the codebase.
